+++ mobile/android/android-components/components/browser/thumbnails/src/main/java/mozilla/components/browser/thumbnails/ThumbnailsMiddleware.kt
@@ -6,6 +6,7 @@ package mozilla.components.browser.thumbnails
 
 import mozilla.components.browser.state.action.BrowserAction
 import mozilla.components.browser.state.action.ContentAction
+import mozilla.components.browser.state.action.TabListAction
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.thumbnails.storage.ThumbnailStorage
 import mozilla.components.lib.state.Middleware
@@ -28,6 +29,23 @@ class ThumbnailsMiddleware(
                 // Store the captured tab screenshot from the EngineView when the session's
                 // thumbnail is updated.
                 thumbnailStorage.saveThumbnail(action.sessionId, action.thumbnail)
+            }
+            is TabListAction.RemoveAllNormalTabsAction -> {
+                store.state.tabs.filterNot { it.content.private }.forEach { tab ->
+                    thumbnailStorage.deleteThumbnail(tab.id)
+                }
+            }
+            is TabListAction.RemoveAllPrivateTabsAction -> {
+                store.state.tabs.filter { it.content.private }.forEach { tab ->
+                    thumbnailStorage.deleteThumbnail(tab.id)
+                }
+            }
+            is TabListAction.RemoveAllTabsAction -> {
+                thumbnailStorage.clearThumbnails()
+            }
+            is TabListAction.RemoveTabAction -> {
+                // Delete the tab screenshot from the storage when the tab is removed.
+                thumbnailStorage.deleteThumbnail(action.tabId)
             }
         }