+++ mobile/android/android-components/docs/changelog.md
@@ -25,6 +25,8 @@ permalink: /changelog/
 
 * **browser-thumbnails**
   * Deletes the tab's thumbnail from the storage when the sessions are removed.
+  * ⚠️ **This is a breaking change**: Removes unused `ThumbnailsUseCases` since we now load thumbnails via `ThumbnailLoader`. See [#7313](https://github.com/mozilla-mobile/android-components/issues/7313).
+  * ⚠️ **This is a breaking change**: Removes `ThumbnailsUseCases` as a parameter in `TabsFeature` and `TabsTrayPresenter`.
 
 # 47.0.0